Remove Text in FAQ modals and put in docs folder
