+++ app/src/components/FaqCard/index.js
@@ -1,11 +1,9 @@
 import React, { useState } from 'react';
 import { View, Text, Image, TouchableOpacity } from 'react-native';
-
 import EmergencyNumbers from '../../components/FAQModals/EmergencyNumbersModal';
 import HelpOfferedModal from '../FAQModals/RecomendationsModal/HelpOfferedModal';
 import HelpRequestModal from '../FAQModals/RecomendationsModal/HelpRequestModal';
 import HowToUseModal from '../FAQModals/HowToUseModal';
-
 import styles from './styles';
 
 export default function FaqCard({ faq }) {
