+++ app/src/docs/FAQ/HelpOfferedRecomendations.js
@@ -1,0 +1,30 @@
+const acceptHelpRequestRecomendationsDescription = `
+##1 Lembre-se que você está lidando com pessoas que precisam de ajuda, logo, tenha paciência, seja prestativo e gentil;
+
+##2 Ao entrar em contato com o ajudado, considere que o mesmo pode estar em alguma situação de risco, por isso, não peça, e nem ofereça, informações adicionais além das apresentadas no app;
+
+##3 Recomendamos, caso seja necessário um encontro, que este ocorra em um local público, e de forma segura.
+`;
+
+const createHelpOfferedRecomendationsDescriptions = `
+##1 Se um objeto for ofertado, certifique-se de que o mesmo está em boas condições de uso;
+
+##2 Certifique-se, ao escolher um ajudado, de que este realmente precisa da ajuda;
+
+##3 Caso seja ofertado um serviço (apoio social, psicológico ou físico, transporte de emergência e pequenos serviços), lembre-se de realizá-lo da melhor forma possível, e sempre com paciência e prestatividade.
+`;
+
+const helpOfferedRecomendations = [
+    {
+        id: '1',
+        title: 'Recomendações importantes para aceitar um pedido de ajuda',
+        description: acceptHelpRequestRecomendationsDescription,
+    },
+    {
+        id: '2',
+        title: 'Recomendações importante para oferecer ajuda',
+        description: createHelpOfferedRecomendationsDescriptions,
+    },
+];
+
+export default helpOfferedRecomendations;
