+++ app/src/docs/FAQ/HelpRequestRecomendations.js
@@ -1,0 +1,31 @@
+const acceptHelpOfferRecomendationsDescription = `
+##1 Verifique se a oferta está de acordo com a sua necessidade, caso contrário deixe a oportunidade para alguém que realmente precise;
+
+##2 Caso a oferta se adeque à sua situação, aceite-a e aguarde o contato do responsável por ela;
+
+##3 Ao receber a ajuda, lembre-se de agradecer ao voluntário.
+`;
+
+const createHelpRequestRecomendationsDescriptions = `
+##1 Lembre-se sempre que você está lidando com outras pessoas, então não peça por coisas impossíveis, nem por dinheiro pois é importante que todos possam ajudar;
+
+##2 Ao receber voluntários para te ajudar, nunca esqueça de ser educado e gentil, pois eles estão tentando fazer o possível para atender à sua solicitação na situação pela qual você está passando;
+
+##3 Recomendamos, caso seja necessário um encontro, que este ocorra em um local público, e de forma segura.
+`;
+
+const helpOfferedRecomendations = [
+    {
+        id: '1',
+        title:
+            'Recomendações importantes para se interessar por alguma oferta de ajuda',
+        description: acceptHelpOfferRecomendationsDescription,
+    },
+    {
+        id: '2',
+        title: 'Recomendações importantes para criar de pedidos de ajuda ',
+        description: createHelpRequestRecomendationsDescriptions,
+    },
+];
+
+export default helpOfferedRecomendations;
