+++ app/src/docs/FAQ/HowToBeVolunteer.js
@@ -1,0 +1,10 @@
+const howToVolunteerRecomendations = [
+    {
+        id: '1',
+        title: 'Como ser voluntário?',
+        description:
+            'Para ser voluntário, é só olhar no mapa da página inicial do app, e clicar no ícone.',
+    },
+];
+
+export default howToVolunteerRecomendations;
