+++ app/src/docs/FAQ/HowToChooseHelpOffered.js
@@ -1,0 +1,10 @@
+const howToChoseHelpOfferRecomendations = [
+    {
+        id: '1',
+        title: 'Como escolher uma oferta de ajuda?',
+        description:
+            'Para escolher uma oferta de ajuda, é só olhar no mapa da página inicial do app, e clicar no ícone de oferecer ajuda',
+    },
+];
+
+export default howToChoseHelpOfferRecomendations;
