+++ app/src/docs/FAQ/HowToCreateHelp.js
@@ -1,0 +1,23 @@
+const HowToCreateHelpRequestModal = `
+##1 Para pedir uma ajuda clique no ícone;
+
+##2 Preencha os dados da página de pedidos: título, categoria e descrição;
+
+##3 Depois de criado o seu pedido, é só esperar que alguém ofereça ajuda;
+
+##4 Quando alguém oferecer ajuda, uma notificação será enviado pelo app;
+
+##5 O contato entre você e o voluntário a lhe ajudar será feito através do telefone ou alguma aplicativo de conversa que vocês utilizem (Exemplo: WhatsApp, Facebook);
+
+##6 Depois que você recebeu a sua ajuda, é só entrar no app e finalizar o seu pedido.
+`;
+
+const createHelpRecomendations = [
+    {
+        id: '1',
+        title: 'Como criar um pedido de ajuda?',
+        description: HowToCreateHelpRequestModal,
+    },
+];
+
+export default createHelpRecomendations;
