+++ app/src/docs/FAQ/faqOptions.js
@@ -2,22 +2,22 @@ const faqOptions = [
     {
         id: 1,
         description: 'Como usar o Mia-Ajuda?',
-        icon: require('../../assets/images/blueCatCard.png'),
+        icon: require('../../../assets/images/blueCatCard.png'),
     },
     {
         id: 2,
         description: 'Recomendações sobre pedidos de ajuda',
-        icon: require('../../assets/images/exclamation.png'),
+        icon: require('../../../assets/images/exclamation.png'),
     },
     {
         id: 3,
         description: 'Recomendações sobre ofertas de ajuda',
-        icon: require('../../assets/images/hand.png'),
+        icon: require('../../../assets/images/hand.png'),
     },
     {
         id: 4,
         description: 'Contatos Importantes',
-        icon: require('../../assets/images/phone.png'),
+        icon: require('../../../assets/images/phone.png'),
     },
 ];
